refactor: remove buildDeferredActionDigest from deferral actions
