+++ account-kit/smart-contracts/src/ma-v2/index.ts
@@ -34,7 +34,7 @@ export {
   getDefaultTimeRangeModuleAddress,
   getDefaultWebauthnValidationModuleAddress,
 } from "./modules/utils.js";
-export { buildFullNonceKey } from "./utils.js";
+export { buildFullNonceKey, buildDeferredActionDigest } from "./utils.js";
 export { allowlistModuleAbi } from "./modules/allowlist-module/abis/allowlistModuleAbi.js";
 export { AllowlistModule } from "./modules/allowlist-module/module.js";
 export { nativeTokenLimitModuleAbi } from "./modules/native-token-limit-module/abis/nativeTokenLimitModuleAbi.js";
